+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Option to show torrent properties in a panel in the main window instead of dialog
+- Ability to set labels on torrents and filter torrent list by labels
 
 ### Changed
 - Options dialog is rearranged to use multiple tabs