+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Option to show torrent properties in a panel in the main window instead of dialog
 - Ability to set labels on torrents and filter torrent list by labels
 - Option to display relative time
+- Option to display only names of download directories in sidebar and torrents list
 
 ### Changed
 - Options dialog is rearranged to use multiple tabs